- terminal gui moved to separate file - main refactor
